Introduce helpers.c and unquote_string().
The unquoting function is the first to be moved to the new file which will
host various utility functions to be shared by multiple files but which do not
fit elsewhere.

unquote_string() extracts a possibly quoted phrase from a string by
using a set of separator characters, which is useful to allow user input
to be optionally quoted to support spaces within names such a region or
file names.